// src/store/movies/interfaces/MoviesTopRated.ts
// Generated by https://quicktype.io

export interface TopRated {
	page: number;
	results: MoviesTopRated[];
	total_pages: number;
	total_results: number;
}

export interface MoviesTopRated {
	adult: boolean;
	backdrop_path: string;
	genre_ids: number[];
	id: number;
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	release_date: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}